removed the unsed files and added the version 2 folder with only used function and folder split up
